# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/content_authors": {
            "get": {
                "description": "Get Content Authors\nGet a list of content authors for which contents exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content Authors",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/content_tags": {
            "get": {
                "description": "Get Content Tags\nGet a list of content tags for which contents exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Content Tags",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/content_titles": {
            "get": {
                "description": "List Content Titles\nList Content Titles by type, author, or tag, paging with reverse, limit, and offset.\nOne of type, author, or tag are required. Optionally, filter results with search terms.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Content"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "enum": [
                            "BOOK",
                            "CHAPTER",
                            "ARTICLE",
                            "CATEGORY"
                        ],
                        "type": "string",
                        "description": "Type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Author",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tag",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search Terms, separated by spaces",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Any Match? (default: false; all search terms must match)",
                        "name": "any",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Sort by Value? (not paginated; default: false)",
                        "name": "sorted",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse Order (default: false)",
                        "name": "reverse",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset (default: forward/reverse alphanumeric)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contents",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/content.Content"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid parameter)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "content.Author": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "content.Content": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/content.Author"
                    }
                },
                "comment": {
                    "type": "string"
                },
                "content": {
                    "$ref": "#/definitions/content.Section"
                },
                "createdAt": {
                    "type": "string"
                },
                "editorId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageCount": {
                    "type": "integer"
                },
                "linkCount": {
                    "type": "integer"
                },
                "sectionCount": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "versionId": {
                    "type": "string"
                },
                "wordCount": {
                    "type": "integer"
                }
            }
        },
...
        "main.APIEvent": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "eventID": {
                    "type": "string"
                },
                "logLevel": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
    }
}